requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.16.0
djangorestframework-simplejwt>=5.5.0
django-cors-headers>=4.9.0
openpyxl>=3.1.0
